--- storehaus-caliper/src/main/scala/com/twitter/storehaus/caliper/Runner.scala
@@ -1,11 +1,0 @@
-package com.twitter.storehaus.caliper
-
-import com.google.caliper.{Runner => CaliperMain}
-
-object Runner {
-
-  def main(args: Array[String]) {
-    CaliperMain.main(args)
-  }
-
-}
