--- storehaus-caliper/src/main/scala/com/twitter/storehaus/caliper/WriteThroughCacheBenchmark.scala
@@ -1,77 +1,0 @@
-package com.twitter.storehaus.caliper
-
-import com.google.caliper.{Param, SimpleBenchmark}
-import com.twitter.algebird.HyperLogLogMonoid
-import com.twitter.bijection._
-import com.twitter.storehaus._
-import com.twitter.conversions.time._
-import com.twitter.algebird._
-import com.twitter.storehaus.cache.{RollOverFrequencyMS, WriteOperationUpdateFrequency,
-  HeavyHittersPercent, MutableCache}
-import scala.math.pow
-import com.twitter.storehaus.algebra._
-import com.twitter.util.{Duration, Timer, Future}
-
-class DelayedStore[K, V](val self: Store[K, V])(implicit timer: Timer) extends StoreProxy[K, V] {
-  override def put(kv: (K, Option[V])): Future[Unit] = {
-    self.put(kv).delayed(10.milliseconds)
-  }
-
-  override def get(kv: K): Future[Option[V]] = {
-    self.get(kv).delayed(10.milliseconds)
-  }
-
-  override def multiGet[K1 <: K](ks: Set[K1]): Map[K1, Future[Option[V]]] =
-    self.multiGet(ks).map { case (k, v) =>
-      (k, v.delayed(10.milliseconds))
-    }
-
-  override def multiPut[K1 <: K](kvs: Map[K1, Option[V]]): Map[K1, Future[Unit]] =
-    self.multiPut(kvs).map { case (k, v) =>
-      (k, v.delayed(10.milliseconds))
-    }
-}
-
-class WriteThroughCacheBenchmark extends SimpleBenchmark {
-  implicit val custTimer = new com.twitter.util.ScheduledThreadPoolTimer(20)
-  import StoreAlgebra._
-
-  implicit val hllMonoid = new HyperLogLogMonoid(14)
-
-  @Param(Array("100", "1000", "10000"))
-  var numInputKeys: Int = 0
-
-  @Param(Array("10000", "100000"))
-  var numElements: Int = 0
-
-  var inputData: Seq[Map[Long, HLL]] = _
-
-  var store: MergeableStore[Long, HLL] = _
-
-  var noCacheStore: MergeableStore[Long, HLL] = _
-
-  override def setUp(): Unit = {
-    val rng = new scala.util.Random(3)
-    val byteEncoder = implicitly[Injection[Long, Array[Byte]]]
-    def size: Int = rng.nextInt(10) + 1 // 1 -> 10
-    def hll(elements: Set[Long]): HLL =
-      hllMonoid.sum(elements.map(byteEncoder(_)).map(hllMonoid.create))
-    val inputIntermediate = (0L until numElements).map {_ =>
-      val setElements = (0 until size).map{_ => rng.nextInt(1000).toLong}.toSet
-      (pow(numInputKeys, rng.nextFloat).toLong, hll(setElements))
-    }.grouped(20)
-
-    inputData = inputIntermediate.map(s => MapAlgebra.sumByKey(s)).toSeq
-
-    val delayedStore = new DelayedStore(new ConcurrentHashMapStore[Long, HLL])
-    val hhStore = HHFilteredStore.buildStore[Long, HLL](
-      new ConcurrentHashMapStore[Long, HLL],
-      MutableCache.ttl(Duration.fromSeconds(10000), numElements),
-      HeavyHittersPercent(0.5f),
-      WriteOperationUpdateFrequency(1),
-      RollOverFrequencyMS(10000000L)
-    )
-    store = new WriteThroughStore(delayedStore, hhStore).toMergeable
-    noCacheStore = delayedStore.toMergeable
-  }
-}
